requirements.txt
langchain 
#==0.3.25
langchain-community 
#==0.3.24
langchain-google-community 
#==2.0.7
langchain-google-vertexai 
#==2.0.24
langchain-chroma 
#==0.2.4
protobuf
#==3.20
pypdf

